upgrade avalanche-platform to v3
