+++ .changeset/smooth-eels-boil.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/avalanche-platform-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
