--- packages/sources/avalanche-platform-test/src/endpoint/balance.ts
@@ -1,43 +1,0 @@
-import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
-import { InputParameters } from '@chainlink/external-adapter-framework/validation'
-import { config } from '../config'
-import { httpTransport } from '../transport/balance'
-
-export const inputParameters = new InputParameters({
-  addresses: {
-    aliases: ['result'],
-    array: true,
-    type: {
-      address: {
-        type: 'string',
-        description: 'an address to get the balance of',
-        required: true,
-      },
-      network: {
-        type: 'string',
-        description: 'the name of the network protocol',
-        default: 'avalanche',
-      },
-    },
-    description:
-      'An array of addresses to get the balances of (as an object with string `address` as an attribute)',
-    required: true,
-  },
-})
-
-export type BaseEndpointTypes = {
-  Parameters: typeof inputParameters.definition
-  Response: {
-    Result: null
-    Data: {
-      result: { balance: string }[]
-    }
-  }
-  Settings: typeof config.settings
-}
-
-export const endpoint = new AdapterEndpoint({
-  name: 'balance',
-  transport: httpTransport,
-  inputParameters,
-})
