--- packages/sources/avalanche-platform/schemas/env.json
@@ -1,19 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/avalanche-platform-adapter.json",
-  "title": "Avalanche Platform Adapter",
-  "description": "Chainlink External adapter for reading from the Avalanche Platform chain's API",
-  "required": ["P_CHAIN_RPC_URL"],
-  "type": "object",
-  "properties": {
-    "P_CHAIN_RPC_URL": {
-      "type": "string",
-      "description": "Full RPC URL for the avalanche platform chain (e.g. https://api.avax.network/ext/bc/P)",
-      "required": true
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
