+++ packages/sources/avalanche-platform/test/integration/adapter.test.ts
@@ -1,48 +1,49 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import { SuperTest, Test } from 'supertest'
-import { server as startServer } from '../../src'
+import {
+  TestAdapter,
+  setEnvVariables,
+} from '@chainlink/external-adapter-framework/util/testing-utils'
+import * as nock from 'nock'
 import { mockBalanceSuccess } from './fixtures'
-import { setupExternalAdapterTest, SuiteContext } from '@chainlink/ea-test-helpers'
 
 describe('execute', () => {
-  const id = '1'
-  const context: SuiteContext = {
-    req: null,
-    server: startServer,
-  }
+  let spy: jest.SpyInstance
+  let testAdapter: TestAdapter
+  let oldEnv: NodeJS.ProcessEnv
 
-  const envVariables = {
-    METRICS_ENABLED: 'false',
-    CACHE_ENABLED: 'false',
-    P_CHAIN_RPC_URL: 'http://localhost:3500/ext/bc/P',
-  }
+  beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    process.env.P_CHAIN_RPC_URL = process.env.P_CHAIN_RPC_URL ?? 'http://localhost:3500/ext/bc/P'
+    const mockDate = new Date('2001-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
 
-  setupExternalAdapterTest(envVariables, context)
+    const adapter = (await import('./../../src')).adapter
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      testAdapter: {} as TestAdapter<never>,
+    })
+  })
 
-  describe('balance api', () => {
-    it('(stakeAmount) should return success', async () => {
-      const data: AdapterRequest = {
-        id,
-        data: {
-          result: [
-            {
-              address: 'P-fuji1vd9sddlllrlk9fvj9lhntpw8t00lmvtnqkl2jt',
-              network: 'avalanche-fuji',
-            },
-          ],
-        },
+  afterAll(async () => {
+    setEnvVariables(oldEnv)
+    await testAdapter.api.close()
+    nock.restore()
+    nock.cleanAll()
+    spy.mockRestore()
+  })
+
+  describe('balance endpoint', () => {
+    it('should return success', async () => {
+      const data = {
+        result: [
+          {
+            address: 'P-fuji1vd9sddlllrlk9fvj9lhntpw8t00lmvtnqkl2jt',
+            network: 'avalanche-fuji',
+          },
+        ],
       }
-
       mockBalanceSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 })
