--- packages/sources/avalanche-platform/test/unit/balance.test.ts
@@ -1,36 +1,0 @@
-import { AdapterError, Requester } from '@chainlink/ea-bootstrap'
-import { assertError } from '@chainlink/ea-test-helpers'
-import { makeExecute } from '../../src'
-
-describe('execute', () => {
-  const jobID = '1'
-  const execute = makeExecute()
-
-  process.env.P_CHAIN_RPC_URL = process.env.P_CHAIN_RPC_URL || 'fake_rpc_url'
-
-  describe('validation error', () => {
-    const requests = [
-      { name: 'empty body', testData: {} },
-      { name: 'empty data', testData: { data: {} } },
-      {
-        name: 'empty addresses',
-        testData: { id: jobID, data: { addresses: [] } },
-      },
-      {
-        name: 'empty result',
-        testData: { id: jobID, data: { result: [] } },
-      },
-    ]
-
-    requests.forEach((req) => {
-      it(`${req.name}`, async () => {
-        try {
-          await execute(req.testData as never, {})
-        } catch (error) {
-          const errorResp = Requester.errored(jobID, error as AdapterError)
-          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
-        }
-      })
-    })
-  })
-})
